+++ src/ParameterObject.cpp
@@ -40,7 +40,7 @@ namespace eipScanner {
 	}
 
 	ParameterObject::ParameterObject(cip::CipUint instanceId, bool fullAttributes, size_t typeSize)
-		: _instanceId(instanceId)
+		:  BaseObject(CLASS_ID, instanceId)
 		, _hasFullAttributes(fullAttributes)
 		, _value(typeSize)
 		, _maxValue(typeSize)
@@ -53,21 +53,21 @@ namespace eipScanner {
 		, _scalingOffset(0) {
 	}
 
-	ParameterObject::ParameterObject(cip::CipUint id, bool fullAttributes,
+	ParameterObject::ParameterObject(cip::CipUint instanceId, bool fullAttributes,
 			const SessionInfo::SPtr &si,
 			const MessageRouter::SPtr& messageRouter)
-		: _instanceId{id}
+		:  BaseObject(CLASS_ID, instanceId)
 		, _name{""}
 		, _hasFullAttributes{fullAttributes}
 		, _isScalable{false}
 		, _messageRouter{messageRouter} {
 
 
-		Logger(LogLevel::DEBUG) << "Read data from parameter ID=" << id;
+		Logger(LogLevel::DEBUG) << "Read data from parameter ID=" << instanceId;
 
 		auto response = _messageRouter->sendRequest(si,
 				ServiceCodes::GET_ATTRIBUTE_SINGLE,
-				EPath(CLASS_ID, _instanceId,
+				EPath(CLASS_ID, instanceId,
 						ParameterObjectAttributeIds::DATA_SIZE),{});
 
 
@@ -86,7 +86,7 @@ namespace eipScanner {
 		}
 
 		response = messageRouter->sendRequest(si,
-				cip::ServiceCodes::GET_ATTRIBUTE_ALL, cip::EPath(CLASS_ID, _instanceId), {});
+				cip::ServiceCodes::GET_ATTRIBUTE_ALL, cip::EPath(CLASS_ID, instanceId), {});
 
 		if (response.getGeneralStatusCode() == GeneralStatusCodes::SUCCESS) {
 			Buffer buffer(response.getData());
@@ -101,7 +101,7 @@ namespace eipScanner {
 			CipWord descriptor;
 			buffer >> descriptor >> reinterpret_cast<CipUsint&>(_type);
 
-			Logger(LogLevel::DEBUG) << "Parameter object ID=" << _instanceId
+			Logger(LogLevel::DEBUG) << "Parameter object ID=" << instanceId
 									<< " has descriptor=0x" << std::hex << descriptor
 									<< " scalable=" << _isScalable;
 
@@ -128,7 +128,7 @@ namespace eipScanner {
 						 ++attrId) {
 						auto response = _messageRouter->sendRequest(si,
 															   ServiceCodes::GET_ATTRIBUTE_SINGLE,
-															   EPath(CLASS_ID, _instanceId, attrId),
+															   EPath(CLASS_ID, instanceId, attrId),
 															   {});
 
 						if (response.getGeneralStatusCode() != GeneralStatusCodes::SUCCESS) {
@@ -149,8 +149,8 @@ namespace eipScanner {
 				std::runtime_error("Not enough data in the response");
 			}
 
-			Logger(utils::DEBUG) << "Read Parameter Object"
-				<< " ID=" << id
+			Logger(LogLevel::DEBUG) << "Read Parameter Object"
+				<< " ID=" << instanceId
 				<< " ValueSize=" << _value.size()
 				<< " ValueType=0x" << std::hex << static_cast<int>(_type)
 				<< " Name=" << _name;
@@ -163,7 +163,7 @@ namespace eipScanner {
 	void ParameterObject::updateValue(const SessionInfo::SPtr& si) {
 		auto response = _messageRouter->sendRequest(si,
 								ServiceCodes::GET_ATTRIBUTE_SINGLE,
-								EPath(CLASS_ID, _instanceId,
+								EPath(CLASS_ID, getInstanceId(),
 								ParameterObjectAttributeIds::VALUE),{});
 
 		if (response.getGeneralStatusCode() == GeneralStatusCodes::SUCCESS) {
@@ -183,10 +183,6 @@ namespace eipScanner {
 
 	cip::CipDataTypes ParameterObject::getType() const {
 		return _type;
-	}
-
-	cip::CipUint ParameterObject::getInstanceId() const {
-		return _instanceId;
 	}
 
 	bool ParameterObject::hasFullAttributes() const {