+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# Release 1.1.0 
+
+**Features**:
+
+* add Identity Object #5, @flipback
+* add utils::LogAppender to customise logging #5, @flipback
+
 # Release 1.0.0 (2019-12-08)
 
 Initial release